test(radio-group): migrate
